Patched stupid buffer segmentation fault
Apparently, the buffer that the generated code is supposed to be put in was not being initialized properly. For some very stupid reason, x86 CPUs (at least roestta) have a very weird behavior when it initializes, the previous `len` value was left at a random value and was not cleared. And ta know, when you try to write to 12930128310 or something in an empty buffer, you would get a segmentation fault.