Changed type (#8)
* Changed type

Changed typcast for offsets from `uint8_t`s to a signed `int8_t`
as well as the signed double-worded counterpats, as specified in
commit c78baced5e91160be835d4c90204c5eb10fb0ab0 where the offset
type was changed to a signed value. (Specified in the Intel man-
ual as well!)

---

Note - The 4 byte displacement offset is NOT typed as `uint32_t`
or any 4 byte-sized data type due to the `buf_write` function
requiring an array of `uint8_t`s. However, they have been bunched
together in a int32_t value with is abstracted as a unsigned char
array?!?!?

* Corected commit

In commit number 96cce7c232baf6f3c7d9d52801dad58def9fc1a9, the chan-
ges specified in the following statement of the commit description:

> Note - The 4 byte displacement offset is NOT typed as uint32_t or
> any 4 byte-sized data type due to the buf_write function requiring
> an array of uint8_ts. However, they have been bunched together in
> a int32_t value with is abstracted as an unsigned char array?!?!?

Was not included in the previous change.

Now, this change **has** been included as outlined, well, *hopefu-
lly*