Added macro to testing framework - `assert_eq_buf`
Added the new `assert_eq_buf` macro to the testing framework in
`test.h` this macro is used in unit tests and can check if two
buffers, `a` and `b` have the same length and values data pairs.

This macro is especially useful when unit tests are rolled out for
indivdual instructons to compare buffers and expected outcomes.