Made variables private
This change made the previosuly global `label_table` and the other
`lable_table_size` variables static and added "getter" (Well, C do-
es not have getters) functions that *gets* the values of the label
table values and it's size.

So, having the `label_table` and `label_table_size` variables hid-
den away prevents mistakes by both the caller and internal functions
that may accidentally: 1. redefine or attempt to clash with the
variable names, and 2. unintentionally modify values of the label
table's values. (So, basically, you can't be stupid enough to even
attempt to modify a function's value)