Renamed enums
This commit renamed all the enocder identitiy enums from the prev-
ious `OP_` prefix to the `ENC_` prefix. Previously, the enocder
identities were placed in the `operand.h` file and held the `OP_`
prefix before it was moved to this file. However despite being moved
to another module, the old name stayed around until it was finally
renamed in this commit.