Renamed `MR` and `RM` common function
Previous name of the common function for the `MR` and `MR` encoder
identities carried the un-descripitive and in conclusive name of
 `mr_rm_ref()`, this new name - `rm_mr_common` should better explain
the funcitonality of this encoder helper function.

Further more, this commit also removed the arbitary boolean param-
eter value that described if the function was used to encode the
`RM` or `MR` encoder identity, now this function simply takes a
enumeration of the current identity and changes behavior based on
this.