Renamed `enc_lookup`'s definition to match source
As per the commit message in 64b99a57e72d24a81717cd49e88d8ca07860e636:

> - The `instr_encode_func` and `instr_encoder_t` function and type
> respectively has been moved over to the `encoder.h` file along with
> the renaming of their prefixes to `enc` (Which stands for encoder)
> rather than the old `instr` prefix (aka *instruction*)

For somereason this was not carried over for the header definiton,
causing linkage errors during linktime. This commit just matches
this with the function prototype in the source file. (No more linker
issues now :))