Added dummy instruction enum
Previously, the assembler used the final instruction in the enum
as a "indicator" for when the direcives start. This brings out a
very very obvious issue where if a developer fails to change the
`IS_DIRECTIVE` macro can make the assembler think that a instruction
that is valid as a directive. Therefore, a new "dummy" has been added
and all directives must be **after** the dummy