Revised `op_write_opcode` function
Happy new year, this pull has revised the `op_write_opcode` function,
which was previously mis-interpreted. According to the Intel x86 docs,
a instruction **including** a byte-sized operand **MUST** use the byte
instruction opcode. This is demonstrated in the MOVZX instruction's docs
on 4-136 Vol. 2B where a seperate opcode is used with `r/m8`.