Added `has_byte_opcode` member to every table (#28)
Since the introduction of the `has_byte_opcode` member in the
`instr_encode_table` struct in #21, basically no instruction encode
table *actually* uses this member of the struct; casuing mis-
interpretation and usage of this member in other functions
and modules. This pull has added this boolean value to *most*
instruction encoder tables, in order to support functions and
conditionals that depend of this member exisiting.