Fixed encoder clash
This pull has fixed the strange encoders clash between the O and I
encoder functions, previously the encoders O and I were both to be
thought as one `imm` (immediate value) operand on the first operand
location.

Upon futher review, it was found that the I operand encoder function
identity consists of a `imm` operand **exclusively** and does not
contain any more oprands. Contary to the I identity, the O identity
**ALSO** contains a **accumulator** as the first operand and a `imm`
**following** the **accumulator** instead.