Fixed error message (#33)
In the previous pull - #32 where the `pre_movzx` pre-processor 
was renamed to `pre_small_operand_sizes`, the error message 
was not changed and remained to shown as a "MOVZX" instruction
fault as per the patch. This pull has corrected the error message to 
point to both the MOVZX and MOVSX instructions instead.