Added case for `NULL` instructions
This pull has addded a "edge case" for the `INSTR_NULL` cases that may
be passed into the `instr_get_tab` funciton. Further testing has found
that this function may be problematic and force segmentation faults
when faced with a `NULL` value in either the operands or `instr`
member. Now, this function will intentionally skip any `NULL` or `INST-
R_NULL` values.