`instr_gen()` fixes and improvements (#37)
* Added `instr_gen` function

This pull will add the `instr_gen()` function into the Jas assembler,
this function will be a wrapper for the previous `instruction_t` struct
in representing instructions and allowing easier definition of instructions
that is one level lower than the textual format, allowing it to be def-
ined pragmatically.

* Modified `instr_gen` function

This commit has modified the `instr_gen` function to leverage the
previously instroduced `op_construct_operand` function. Now, this
function will use variable arguments and group 3 arguments to be
passed onto the `op_construct_operand()` function to be constructed.

* Added declaration and documentation

This commit has added the delaration and docs for the `instr_gen`
function added to the source tree in the previous commit, please
see commit 39845ff6190a6f755dd2a557ecc6a0db23058924 for more info.

* Chnaged argument size

As per the `op_construct_operand` function documentation, the x64
and x86 `offset` will take `size_t` or 64-bit sized offsets instead
of the previous `uint8_t` sized.

* Finished `instr_gen` funciton

This commit has finally finalised the funcitonality of the instruction
generation function, in particular some changes of the function can
be affected at this stage:

  - Operands are put in order of type, data, offset respectively
  - Labels are only allowed when using the `rel` operand type
  - Therre is no need to allocate memory or provide a pointer,
    a value can be directly passed in.

Although these features has already been implemented, this function
still lacks significant documentation, which will be worked on in the
near future.