Fixed issue with register not assigning to operand
For some completely random and obscure reason, when using the `-O3`
or whatever optimisation other than the default, the `instruction.c`
unit test testing against the `instr_gen` function breaks when called,
further digging has revealed for some reason, the `data` value is not
being assigned with the required value. (and instead holds a completely
insane in place).

This pull has combined the two statements into a compound literal, red-
ucing complexity to the code which prevents it from being "optimised"
out of the program and therefore breaking whenever a optimisation flag
is set in the compiler options.