Renamed `INSTR_TERMINATOR` to `INSTR_TAB_NULL` (#41)
This commit has changed the name of the previous terminator to a
"null wrapper" for a more "practical" name. For example, the `instr_
get_tab` function uses the `INSTR_TERMINATOR` as a default empty
value, which does not fall under the "termination" category and does
not need to be listed as that.