Fixed incorrect offset values. (#43)
After the switch to updating the ELF offset using pointers to allow
the offset value to be updated automatically within the function.
Some other functionality has broke since the changes in #39 and as
outlined above. First of all, the outstanding issue causing issues
after the change is the "pre mature" modification of the offset
before it was used.

In particular, the modification line:
  > *off += sect_sz;

Was placed **before** the line where *it was used*:
  > buf_write(&ret, (uint8_t *)off, 8);

This means, the value is incorrectly referenced, however, this is
not the primary issue *(I think)*. The primary issue I think is
the fact that I am referencing the pointer to the `off` pointer.

Previously, the offset was written as `&off` which is the **address**
of `off`, instead of the value.