Fix instr gen (#50)
* Fixed `instr_gen` function

After some days of debugging later, (well, it was very very confusing)
the root issue for the incorrect instruction was finally found.

Since all local variables in C included within a funciton is stored on
the stack frame, it will get cleaned-up after the function exits/returns,
therefore, when `instruction_t` struct was returned from the `instr_gen`
function, we were reading from the *stack frame* instead of *actual
memory*. After other functions were called, the stack frame where the
previous `instruction_t` struct was stored was overwitten ang and placed
back in with some random garbage.

This pull is a *almost stub* implementation/fix for the `instr_gen` function,
in particular, it allows memory to be allocated dynamically to support the
storage of these titems that were temporary when implemented on the stack
frame.

---

**Unrelated:** This commit has also modified the names of the `div` instruction
enocder table as per the changes explained in #49.

* Encapsulated code into a do-while loop

Previously, when the `alloc_operand_data` macro was left "freestanding",
there was no protection of the macro being confused with other syntax.
If there are more complicated bodies of code above and below this macro,
concerns may arise when mangling and clashing against eachother.