Instr generator funcs fix (#55)
* Added dynamic memory allocations to generators

After the fix in #50, which only covered the `instr_gen` function,
this pull has taken a similar method in patching the other generation
function for generating a `instruction_t` struct dynamically.

* Exposed `label.h` to public API

* **unrelated** Used `INSTR_DIRECTIVE` instead to cover directives

After the introduction of the `instr_write_bytes` function and the
`INSTR_DIR_WRT_BUF` directive, the `codegen` function started to allow
the same as the `dd` or `db` directive in nasm. However, the function
caused the program to de-ref some of variables and  started to segfault
after attemping to find a corrisponding instruction encoder table (Which
does not exist)

This commit has used this directive/macro to filter out and eliminate
the possibility of these instructions being attempted to be worked on

* Added dynamic memory allocations - **stub**

For some reason, basically **everything** inside a function is stored
onto the stack which is not preserved as documented in #50, forcing
everything to be allocated dynamically. Even though this code lacks
the corrispondng `free()` function call, a subsquent pull will be
produced to patch up the issue