Modifications to enocder lookups
This pull has addressed some long-term issues with using a "free
for all" unordered map that frequenty find issues with incorrect
enocdings and *un-optimised* enocder identies and previously needing
a if statement to manually correct for it after the operation.

After this change, **everything** is stored in a single *`multimap`*,
given a instruction reference passed into the function, the implemented
`op_ident_identify` function will rule out any edge cases that are not
shown inside the provided table, reducing chances of a mis-hap.