Added filler element for `enum instructions`
In the previous code (as shown in the patch file), the `instr_get_tab`
function attemps to compare the `instr.instr` enumeration to the
constance `NULL` value. Since the `NULL` value equates to the `INSTR_MOV`
enum, it may cause incorrect readings and a compiler error as well 😝.

This pull has added a `INSTR_NOTHING` value to the `instructions` enum
to seperate values and allow debugging processes to directly and clearly
identify if an instruction is invalid, instead of defaulting to `MOV`.