Fixed type cast issue (#60)
In this pull, a previously not dereferenced pointer has been forced
and casted into the `reg_needs_rex` funciton as shown below:
> !reg_needs_rex((enum registers)op_arr[0].data)

This is immediately incorrect as a memory address may **not** just
magically turn into a register, not without a de-reference to the
pointer. This is further pronounced when we see that the function
**itself** does not allow a pointer to be passed, so having a pointer
forced into there is **very very very dangerous** (This is why we have
rust lol 🦀).