Passed correct `operand_t` typed structure
Previously, `op_modrm_mode()` function used a `op_r` macro (as
defineda and documented in `operand.h`) to determine if a indirect/
direct register operand is used. However, this was incorrectly
invoked and caused a compiler warning, which was ignored. This pull
has changed the register value to the *correct* operand.