Op modrm mode fix (#63)
* Updated macro

This commit has updated the `alloc_operand_data` macro to allow it
to safely allocate enough data when "popping" an argument from the
stack for. Previously, the `va_arg` macro would promote some data
sizes as `int`s and messing up other data types. This commit has
just added a little "one size fits all" solution by allocating a
64-bit sized parameter.

* Passed correct `operand_t` typed structure

Previously, `op_modrm_mode()` function used a `op_r` macro (as
defineda and documented in `operand.h`) to determine if a indirect/
direct register operand is used. However, this was incorrectly
invoked and caused a compiler warning, which was ignored. This pull
has changed the register value to the *correct* operand.