Added support for IN and OUT instructions
This pull has added the support for the IN and OUT instructions
in the Jas assembler, due to the unusual patterns of the encoding
as documented in the Intel documentation, this instruction has been
put off, and I've procrastinated for quite a while of adding the
support for this instruction. Inparticular, instruction has non-
standard operand encodin identity formats where the ZO identity
was used to depict the use of the DX and AL *registers*, which
**had operands**. (See the extact from below)

> Instruction	Op/En		64-Bit Mode
> EF		OUT DX, EAX	ZO

Since the in-consistency, a new encoder identity was implemented
in this pull, namely a `ENC_IGN` idenity that is used as a little
*default* value *other than* the `NULL` value, as seen in previous
commits. This encoder identity will **ignore** all operands and
just encode the operand, passing jobs onto other enocders which;
matches the needs and requirements of the IN and OUT instructions'
encodings.