Refactored and simplified mode-validation functionality
(I'm back) This commit has refactored the mode-validation functions
which was very messy. Previously, the modes had a couple enums with
vaguely documented functions and usages; one in particular which was
the `mode_support` enum (Or something along the lines) described the
modes an instuction can be encoded in was barely used by the assembler.
All (expect one) were allowed all modes for encoding.