Removed `instr_index` field of the label struct + more (#68)
This pull contains several fixes and changes to the assembler,
specifically mostly to the label model:

First, this change has completely removed the dependence and
usage of the `instr_index` feild of the `label` structure, which
was previously rarely used or utilized, or for the matter of fact
ever had any use.

Also, when I was playing with the label generation and code assembling,
some memory management bugs were found, in particular of similar issue
to #57, where the stack-allocated data was being overwirtten by another
function, which *definiatly* was not portable or can be relased. This
issue has since been patched by allocating memory on the heap and refe-
rencing that instead.