Make assembler labels 'first class symbols'
This pull is aimed as reducing the previous 'monopoly' of the label
module where previously, the labels are stored ina global variable
and tracked that way using certain utility function. This solution
is fine, but causes *few* problems:

a) **Thread safety**, If we wanted to multithread this or use this in
specific applications where somehow multihtreading or race-conditions
can potentially be critical, this dis-qualifies it as "safe"

b) **Flexibility**, having everything (labels and symbols) being cooped
up into one little "cubby" prevents the user and dis-allows control, un-
less if they come into the source code and get their hands dirty, which
in my opinion is un-disierable, for both programmers and users, where
wrappers and everything has to be strangely "pushed into eachother".

c) **Maintance & re-useability**. similar in terms of *flexibility*, having
more flexibility allows the module to be packaged more snuggly and allow
it to be re-used in other functions or applications, if needed (later down
the road)

---

This pull achomplish this by having many functions pass-in a `label_table`
and corrisponding size to do the operations on. Also, older software must
also be altered to support dependencies of this new function.