Make assembler labels 'first class symbols' (#73)
* Make assembler labels 'first class symbols'

This pull is aimed as reducing the previous 'monopoly' of the label
module where previously, the labels are stored ina global variable
and tracked that way using certain utility function. This solution
is fine, but causes *few* problems:

a) **Thread safety**, If we wanted to multithread this or use this in
specific applications where somehow multihtreading or race-conditions
can potentially be critical, this dis-qualifies it as "safe"

b) **Flexibility**, having everything (labels and symbols) being cooped
up into one little "cubby" prevents the user and dis-allows control, un-
less if they come into the source code and get their hands dirty, which
in my opinion is un-disierable, for both programmers and users, where
wrappers and everything has to be strangely "pushed into eachother".

c) **Maintance & re-useability**. similar in terms of *flexibility*, having
more flexibility allows the module to be packaged more snuggly and allow
it to be re-used in other functions or applications, if needed (later down
the road)

---

This pull achomplish this by having many functions pass-in a `label_table`
and corrisponding size to do the operations on. Also, older software must
also be altered to support dependencies of this new function.

* Integrated label tables into dependent code.

After many major changes to the label module, we should *in theory*
finally be thread-safe (for now at least, until something big crushes
this), Some aspects dependent on the `label` module has been modified
and adapted to support this change, which were documented in the #73
PR and commit description of 3384fbff3ed9b4907b098fbfc5c9a83a8bae19ad.

Here are some adaptation changed as of this commit:
  - Encoders and the corrisponding `encoder_t` function signatures
    as well as implementations now have two extra types - namely
    the `label_table` and `label_table_size` `label_t` and `size_t`
    values to allow passing of these down into the encoder without
    sharing a single global variable. (Like we used to do)
  - Components of the ELF file generation module which depended on
    the label table's size and relavant information has received a
    similar update like above (where applicable) where information
    is passed down, in abscense of a global value or variable.
    (Note conflicts regarding cross include conflicts and cross-dep-
    endent issues has been found to be problematic in some headers,
    leading to forward-declared items and types)

**Note:** A mistake regarding the mis-dereference of the `label_table`
pointer in `label.c:50` has been found, attempting to reallocate a
pointer which pointed to the actual array, this has been fixed by lit-
erally adding in a de-reference and **should've been caught previously**