Remove strdup for malloc (#74)
* Experimental: Utilise `malloc` instead of `strdup`

There have been portability issues associated with the (although C
standard) `strdup` funciton, which apparently is un-defined on linux
systems. This commit has opted to use the standard `malloc` and `strcpy`
it over to the destination.

* Removed `strdup` call of `instr_gen`

I forgot about this one -  The `instr_gen` function also uses the
`strdup` function to duplicate and allocate memory for the label
string