Changed `has_byte_opcode` into size field (#75)
Previously, the `has_byte_opcode` feild has been added to the
`instr_encode_table` to show wether a instruction encoder table
entry contains a byte opcode (An opcode used **only** when solely
byte operands has been supplied). However, some of the instructions
have inter-modal opcode sizes, for example the byte opcode can be
2 bytes long and where the standard opcode is only 1 byte long.

Previous efforts of adding a `0x90` to "pad" the opcode was used
and retained the same functions of also just simple writing nothing.
This pull has omitted this issue by changing the `has_byte_opcode`
struct member as a `byte_opcode_size` field to allow it to indicate
the pressence of a byte opcode **as well as** a size **at the same
time** (Where a `byte_opcode_size` of > 0 and =< 3 means a present
byte opcode and shall be encoded by the corrisponding encoder fun-
ction)