Added text-based instruction table support
Previously and also as of the current codebase, everything (all the
instruction enocder tables) have been represented in C and in a huge
file of structs. (Not very maintainable or readable, frankly) This
pull has added a compiler for a text-based format (`.tbl` files which
the linux kernel also uses for kernel-system-calls).

In a nutshell, the `.tbl` file has options consisting of the corrisponding
values in a equivilent C structure such as opcode, opcode extention and pre
functions, with some changed. These values will be read and parsed, then
produced into a C syntax to be fed into the compiler as C code. (For more
information, please see the `instruction.tbl` file's comments)