Added makefile target for `tabs.c`
During this migration process, the goal is to (one day) migrate all
the instruction encoder tables to the `instructions.tbl` file to be
compiled by the compiler script and not to be touching any C at all.
One component is the automated script (which is not attached to any-
thing big at the momment as its yet to be completed), that builds
the `tabs.c` file automatically, without direct access and mangling
with the ugly C syntax (ughh I'm traumatized).