Added brackets around `NULL` + script changes
In the compiler script for the text-based instruction enocder bank,
the C compiler has encountered issues where when the array was gen-
erated as `NULL` pruely, it reconised the following elements as the
*members* of the array rather than sibling elements *inside the object*.
The exclusion of quotes in the empty array has also lead to the
throwing of the error: "Invalid integer conversion resulting from
truncation", which supports the relavant assumption that the struct
members has been parsed as array members instead.

Also, a guard clause has also been added to check if the grouping of
the byte opcode, writing in `0` as the byte opcode size if its set as
`NULL`. Previously the compiler has mis-read it has `1` since it counted
the number of commas, defaulting to `1` as the default opcode size.
This worked in the *normal opcode* where it was required but did not
work as intended when put into use in the byte opcode field.