Merge pull request #86 from cheng-alvin/label_free_fix
In the `label_free()` function, the function attempted to
call `free` on the basis of the pointer to the allocated
memory, rather than the actually memory that had to be freed.
Ultimately, causing the program to throw a *"pointer freed
has not been allocated"* error during run- time. This issue
arose since I have attempted to use the pointer of a pointer
to the allocated data to allow the pointer to updated by ref-
erence, making it more seemless. But, this ended in this sh*t.