Fixed `syscall` instruction opcode
`syscall`'s opcode was not inputted correctly during time of writing,
instead of the intended `0x0f, 0x05` opcode, only `0x0f` was entered
as the opcode in the `instructions.tbl` file. This could possibly be
an issue with a direct result of the `compile.js` scripts in #74,
and will be recorded to see if an issue is relating to this script,
rather than human-error.