Encoder bank fixes (#87)
* Corrected `sub` instruction encoder bank (see)

Corrected opcode ext number

During the migration of the sub instruction of #74, for some reason,
the **correct** opcode extension value of `5` was copied over as `8`.
(Possible due to the similarities in writing). However, further digging
around in the Intel manual have revealed the MI idenity carried an
opcode extension of `5` rather than `8`

Furthermore, the byte and normal opcodes have also been pre-
viously mized up, this commit has also corrected that.

* Fixed `syscall` instruction opcode

`syscall`'s opcode was not inputted correctly during time of writing,
instead of the intended `0x0f, 0x05` opcode, only `0x0f` was entered
as the opcode in the `instructions.tbl` file. This could possibly be
an issue with a direct result of the `compile.js` scripts in #74,
and will be recorded to see if an issue is relating to this script,
rather than human-error.