Fixed build system
Previously, there have been some issues regarding the build system,
which has caused build errors and issue with just simply running
`make`. After the addition of the `instructions.tbl` file, an extra
compilation step is required to allow the "raw" text-based file to
be compiled down to a C struct that can be read by the compiler.

Now, the build step has been split up into the `tabs.c` step as well
as a build step for the other code into binary. The `tabs.c` target
is used as pre-requirement of the other code (Since the `instrucion.c`
package for instance includes it and is dependent on it).

This allows for more clarity of what is going on behind the scenes,
previously this was done under the hood, which caused a race-condition
from gcc to compile the `tabs.c` file while everything else was being
compiled. By spliting it up into another build step, there is no
way for it be mixed up during mutlithreading applications.