Reworked ELF generation + changes to ELF generation
This pull has done an overhaul of the ELF generation functionality
of the Jas assembler library. Previously, all ELF object generation
behavior has been baked into the `codegen` function, offering very
little flexibilty for change. This also has complicated the `codegen`
function by not only having more verbose arguments, but also cluttered
in code. This pull has resolved most issues, re-visiting the previously
rushed funcitonality. These changes include:

  - Changes to return type from the `codegen` function of `codegen_ret`
    which exposed access of the label table and label table size members
    to callers. (Something which previously was not allowed)

  - Cleaned up inconsistencies in the `exe.c` file, especially `buf_write`
    calls that had inconsistent endian conversions (some values used the
    jas `endian()` funciton while some were hard-coded, usually without
    much of explaination or comments), creating confusion inside the code.

  - Refactored out type-specific if-checks and hard-coded values: some
    section headers such as the one of the `.symtab` involved inter-dependent
    sections which were mentioned/cited in the section header. Previously,
    these were hard-coded and were of the function's descretion. Now, this
    function has now passed down by the caller through new parameters.

  - *And obviously, the migration of the object generation function*. This
    pull has also seen the refactoring of many things among others, steming
    out from *"reworking the ELF generation"* , which this did infact complete.

This is not to be confused with later integration of the relocation tables
and `extern` functions, this is only a small refactor/overhaul of old software,
infavor of flexibility and maintainablity