Pass by reference rather than value
This commit has patched a issue experienced by #89 where some values
passed in by `label_sect` in the `exe_sect_header()` function is being
passed to a buffer write call where a pointer is expected. This has
lead to the `buf_write` function to de-reference a `NULL` pointer when
nothing is passed in.