Reworked ELF generation + changes to ELF (#89)
* Reworked ELF generation + changes to ELF generation

This pull has done an overhaul of the ELF generation functionality
of the Jas assembler library. Previously, all ELF object generation
behavior has been baked into the `codegen` function, offering very
little flexibilty for change. This also has complicated the `codegen`
function by not only having more verbose arguments, but also cluttered
in code. This pull has resolved most issues, re-visiting the previously
rushed funcitonality. These changes include:

  - Changes to return type from the `codegen` function of `codegen_ret`
    which exposed access of the label table and label table size members
    to callers. (Something which previously was not allowed)

  - Cleaned up inconsistencies in the `exe.c` file, especially `buf_write`
    calls that had inconsistent endian conversions (some values used the
    jas `endian()` funciton while some were hard-coded, usually without
    much of explaination or comments), creating confusion inside the code.

  - Refactored out type-specific if-checks and hard-coded values: some
    section headers such as the one of the `.symtab` involved inter-dependent
    sections which were mentioned/cited in the section header. Previously,
    these were hard-coded and were of the function's descretion. Now, this
    function has now passed down by the caller through new parameters.

  - *And obviously, the migration of the object generation function*. This
    pull has also seen the refactoring of many things among others, steming
    out from *"reworking the ELF generation"* , which this did infact complete.

This is not to be confused with later integration of the relocation tables
and `extern` functions, this is only a small refactor/overhaul of old software,
infavor of flexibility and maintainablity

* Pass by reference rather than value

This commit has patched a issue experienced by #89 where some values
passed in by `label_sect` in the `exe_sect_header()` function is being
passed to a buffer write call where a pointer is expected. This has
lead to the `buf_write` function to de-reference a `NULL` pointer when
nothing is passed in.

* Subsequent patches

Some mistakes were made during the inital commit of this pull, all
were critical and when ran through the `brainfry` program and checked
against the code made critical issues that rendered the object file
as use-less. This commit has resolved such issues with a detailed
breakdown featured below:

- Incorrectly used section header properties to the wrong section
  Since the addition of allowing the caller to decide what to be
  written for section-variable properties, some were passed incorrectly
  to a section that did not require such information, causing mis-
  interpretation of the functionality of the section, where the header
  intended to meta-data *for* the section itself.

- Corrected content offset
  Previous offsets in the `addr` variable was generic, only setting it
  as `0x40`, not accounting for the extra bytes taken up by the section
  headers and assuming that the content starts right at the end of the
  file header (which is obviously not true)

- Added return statement
  Stupid me has forgotten a return statement for the outputted buffer from
  the `exe_generate` function, rendering the data useless and leaking a bunch
  of memory if used incorrectly, hence this version is *very* unsatable.

Please also note a invalid de-reference was written in code and was
addressed in an earlier commit (see a49aa1de616279002a10a3ab742301700ee54209)